Первое задание без дополнительных
