+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="ru.mikhailskiy.intensiv">
+    package="ru.rakhman.moviefinder">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="ru.mikhailskiy.intensiv.MovieFinderApp"
+        android:name="ru.rakhman.moviefinder.MovieFinderApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="ru.mikhailskiy.intensiv.MainActivity">
+        <activity android:name="ru.rakhman.moviefinder.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
