--- app/src/main/java/ru/mikhailskiy/intensiv/data/MockRepository.kt
@@ -1,19 +1,0 @@
-package ru.mikhailskiy.intensiv.data
-
-object MockRepository {
-
-    fun getMovies(): List<Movie> {
-
-        val moviesList = mutableListOf<Movie>()
-        for (x in 0..10) {
-            val movie = Movie(
-                title = "Spider-Man $x",
-                voteAverage = 10.0 - x
-            )
-            moviesList.add(movie)
-        }
-
-        return moviesList
-    }
-
-}
