--- app/src/main/java/ru/mikhailskiy/intensiv/data/Movie.kt
@@ -1,9 +1,0 @@
-package ru.mikhailskiy.intensiv.data
-
-class Movie(
-    var title: String? = "",
-    var voteAverage: Double = 0.0
-) {
-    val rating: Float
-        get() = voteAverage.div(2).toFloat()
-}
