--- app/src/main/java/ru/mikhailskiy/intensiv/network/MovieApiClient.kt
@@ -1,5 +1,0 @@
-package ru.mikhailskiy.intensiv.network
-
-object MovieApiClient {
-
-}
