--- app/src/main/java/ru/mikhailskiy/intensiv/network/MovieApiInterface.kt
@@ -1,6 +1,0 @@
-package ru.mikhailskiy.intensiv.network
-
-
-interface MovieApiInterface {
-
-}
