--- app/src/main/java/ru/mikhailskiy/intensiv/ui/EditTextExtensions.kt
@@ -1,8 +1,0 @@
-package ru.mikhailskiy.intensiv.ui
-
-import android.text.Editable
-import android.widget.EditText
-import androidx.core.widget.addTextChangedListener
-
-fun EditText.afterTextChanged(action: (s: Editable?) -> Unit) =
-    addTextChangedListener(afterTextChanged = action)
