--- app/src/main/java/ru/mikhailskiy/intensiv/ui/feed/MovieItem.kt
@@ -1,29 +1,0 @@
-package ru.mikhailskiy.intensiv.ui.feed
-
-import com.squareup.picasso.Picasso
-import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
-import com.xwray.groupie.kotlinandroidextensions.Item
-import kotlinx.android.synthetic.main.item_with_text.*
-import ru.mikhailskiy.intensiv.R
-import ru.mikhailskiy.intensiv.data.Movie
-
-class MovieItem(
-    private val content: Movie,
-    private val onClick: (movie: Movie) -> Unit
-) : Item() {
-
-    override fun getLayout() = R.layout.item_with_text
-
-    override fun bind(viewHolder: GroupieViewHolder, position: Int) {
-        viewHolder.description.text = content.title
-        viewHolder.movie_rating.rating = content.rating
-        viewHolder.content.setOnClickListener {
-            onClick.invoke(content)
-        }
-
-        // TODO Получать из модели
-        Picasso.get()
-            .load("https://m.media-amazon.com/images/M/MV5BYTk3MDljOWQtNGI2My00OTEzLTlhYjQtOTQ4ODM2MzUwY2IwXkEyXkFqcGdeQXVyNTIzOTk5ODM@._V1_.jpg")
-            .into(viewHolder.image_preview)
-    }
-}
