+++ app/src/main/java/ru/rakhman/moviefinder/data/Movie.kt
@@ -1,0 +1,39 @@
+package ru.rakhman.moviefinder.data
+
+import com.google.gson.annotations.SerializedName
+
+class Movie(
+    posterPath: String,
+    @SerializedName("adult")
+    var isAdult: Boolean,
+    @SerializedName("overview")
+    var overview: String?,
+    @SerializedName("release_date")
+    var releaseDate: String?,
+    genreIds: List<Int>,
+    @SerializedName("id")
+    var id: Int?,
+    @SerializedName("original_title")
+    var originalTitle: String?,
+    @SerializedName("original_language")
+    var originalLanguage: String?,
+    @SerializedName("title")
+    var title: String?,
+    @SerializedName("backdrop_path")
+    var backdropPath: String?,
+    @SerializedName("popularity")
+    var popularity: Double?,
+    @SerializedName("vote_count")
+    var voteCount: Int?,
+    @SerializedName("video")
+    var video: Boolean?,
+    @SerializedName("vote_average")
+    var voteAverage: Double?
+) {
+    @SerializedName("poster_path")
+    var posterPath: String? = null
+        get() = "https://image.tmdb.org/t/p/w500$field"
+
+    val rating: Float
+        get() = voteAverage!!.div(2).toFloat()
+}
