+++ app/src/main/java/ru/rakhman/moviefinder/data/MoviesResponse.kt
@@ -1,0 +1,12 @@
+package ru.rakhman.moviefinder.data
+
+import com.google.gson.annotations.SerializedName
+
+data class MoviesResponse (
+    var page: Int,
+    var results:List<Movie>,
+    @SerializedName("total_results")
+    var totalResults:Int,
+    @SerializedName("total_pages")
+    var totalPages: Int
+)
