+++ app/src/main/java/ru/rakhman/moviefinder/ui/EditTextExtensions.kt
@@ -1,0 +1,8 @@
+package ru.rakhman.moviefinder.ui
+
+import android.text.Editable
+import android.widget.EditText
+import androidx.core.widget.addTextChangedListener
+
+fun EditText.afterTextChanged(action: (s: Editable?) -> Unit) =
+    addTextChangedListener(afterTextChanged = action)
