+++ app/src/main/java/ru/rakhman/moviefinder/ui/feed/MovieItem.kt
@@ -1,0 +1,29 @@
+package ru.rakhman.moviefinder.ui.feed
+
+import com.squareup.picasso.Picasso
+import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
+import com.xwray.groupie.kotlinandroidextensions.Item
+import kotlinx.android.synthetic.main.item_with_text.*
+import ru.rakhman.moviefinder.R
+import ru.rakhman.moviefinder.data.Movie
+
+class MovieItem(
+    private val content: Movie,
+    private val onClick: (movie: Movie) -> Unit
+) : Item() {
+
+    override fun getLayout() = R.layout.item_with_text
+
+    override fun bind(viewHolder: GroupieViewHolder, position: Int) {
+        viewHolder.description.text = content.title
+        viewHolder.movie_rating.rating = content.rating
+        viewHolder.content.setOnClickListener {
+            onClick.invoke(content)
+        }
+
+        // TODO Получать из модели  x
+        Picasso.get()
+            .load(content.posterPath)
+            .into(viewHolder.image_preview)
+    }
+}
