+++ app/src/main/java/ru/rakhman/moviefinder/ui/profile/ProfileAdapter.kt
@@ -1,0 +1,17 @@
+package ru.rakhman.moviefinder.ui.profile
+
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import ru.rakhman.moviefinder.ui.watchlist.WatchlistFragment
+
+class ProfileAdapter(fragment: Fragment, private val itemsCount: Int) :
+    FragmentStateAdapter(fragment) {
+
+    override fun getItemCount(): Int {
+        return itemsCount
+    }
+
+    override fun createFragment(position: Int): Fragment {
+        return WatchlistFragment.newInstance()
+    }
+}
