+++ app/src/main/java/ru/rakhman/moviefinder/ui/watchlist/MoviePreviewItem.kt
@@ -1,0 +1,26 @@
+package ru.rakhman.moviefinder.ui.watchlist
+
+import com.squareup.picasso.Picasso
+import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
+import com.xwray.groupie.kotlinandroidextensions.Item
+import kotlinx.android.synthetic.main.item_with_text.*
+import ru.rakhman.moviefinder.R
+import ru.rakhman.moviefinder.data.Movie
+
+class MoviePreviewItem(
+    private val content: Movie,
+    private val onClick: (movie: Movie) -> Unit
+) : Item() {
+
+    override fun getLayout() = R.layout.item_small
+
+    override fun bind(viewHolder: GroupieViewHolder, position: Int) {
+        viewHolder.image_preview.setOnClickListener {
+            onClick.invoke(content)
+        }
+        // TODO Получать из модели
+        Picasso.get()
+            .load("https://www.kinopoisk.ru/images/film_big/1143242.jpg")
+            .into(viewHolder.image_preview)
+    }
+}
