+++ app/src/main/res/values/strings.xml
@@ -26,4 +26,8 @@
 
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="placeholder">Пёс Киноман</string>
+    <string name="now_played">Сейчас идёт</string>
+    <string name="tv_show_on_air">Сейчас в эфире</string>
+    <string name="tv_show_latest">Cамые свежие телешоу</string>
+    <string name="tv_show_popular">Популярные сериалы</string>
 </resources>
